requirements.txt
slack-bolt==1.19.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
fastapi==0.112.2
uvicorn==0.30.6